+++ src/Control/Lens/Getter.hs
@@ -57,23 +57,26 @@ module Control.Lens.Getter
   , views
   , use
   , uses
-  , peruse
-  , peruses
 
   -- * Storing Getters
   , ReifiedGetter(..)
   , Gettable
   , Accessor
+
+  -- * Deprecated
+  , peruse
+  , peruses
   ) where
 
 import Control.Lens.Classes
 import Control.Lens.Internal
 import Control.Lens.Internal.Combinators
-import Control.Monad.Reader.Class       as Reader
-import Control.Monad.State.Class        as State
+import Control.Monad.Reader.Class as Reader
+import Control.Monad.State        as State
 
 -- $setup
 -- >>> import Control.Lens
+-- >>> import Data.List.Lens
 
 infixl 8 ^., ^&
 infixl 1 &
@@ -91,6 +94,17 @@ infixr 0 ^$
 --
 -- >>> "hello" & length & succ
 -- 6
+--
+-- This combinator is commonly used when applying multiple lens operations in sequence.
+--
+-- >>> ("hello,"world") & _1.element 0 .~ 'j' & _1.element 4 .~ 'y'
+-- ("jelly","world")
+--
+-- This reads somewhat similar to:
+--
+-- >>> flip execState ("hello","world") $ do _1.element 0 .= 'j'; _1.element 4 .= 'y'
+-- ("jelly","world")
+
 (&) :: a -> (a -> b) -> b
 a & f = f a
 {-# INLINE (&) #-}
@@ -99,7 +113,7 @@ a & f = f a
 --
 -- >>> "hello" ^& length
 -- 5
--- >>> import Data.List.Lens
+--
 -- >>> ("hello","world")^._1^&reverse^?!_head
 -- 'o'
 (^&) :: a -> (a -> b) -> b
@@ -141,6 +155,12 @@ to f g = coerce . g . f
 
 
 -- |
+-- When you see this in a type signature it indicates that you can
+-- pass the function a 'Control.Lens.Type.Lens', 'Getter',
+-- 'Control.Lens.Traversal.Traversal', 'Control.Lens.Fold.Fold',
+-- 'Control.Lens.Projection.Projection', 'Control.Lens.Iso.Iso',
+-- or one of the indexed variants, and it will just \"do the right thing\".
+--
 -- Most 'Getter' combinators are able to be used with both a 'Getter' or a
 -- 'Control.Lens.Fold.Fold' in limited situations, to do so, they need to be
 -- monomorphic in what we are going to extract with 'Const'. To be compatible
@@ -152,6 +172,7 @@ to f g = coerce . g . f
 -- you can pass a 'Control.Lens.Fold.Fold' (or
 -- 'Control.Lens.Traversal.Traversal'), otherwise you can only pass this a
 -- 'Getter' or 'Control.Lens.Type.Lens'.
+--
 type Getting r s t a b = (a -> Accessor r b) -> s -> Accessor r t
 
 -------------------------------------------------------------------------------
@@ -175,8 +196,8 @@ type Getting r s t a b = (a -> Accessor r b) -> s -> Accessor r t
 -- "world"
 --
 --
--- It may be useful to think of 'view' as having one of these more restrictive
--- signatures:
+-- As @views@ is commonly used to access the target of a 'Getter' or obtain a monoidal summary of the targets of a 'Fold',
+-- It may be useful to think of it as having one of these more restrictive signatures:
 --
 -- @
 -- 'view' ::             'Getter' s a             -> s -> a
@@ -185,8 +206,18 @@ type Getting r s t a b = (a -> Accessor r b) -> s -> Accessor r t
 -- 'view' ::             'Control.Lens.Type.Simple' 'Control.Lens.Type.Lens' s a        -> s -> a
 -- 'view' :: 'Monoid' m => 'Control.Lens.Type.Simple' 'Control.Lens.Traversal.Traversal' s m   -> s -> m
 -- @
-view :: Getting a s t a b -> s -> a
-view l = runAccessor# (l Accessor)
+--
+-- In a more general setting, such as when working with a monad transformer stack you can use:
+--
+-- @
+-- 'view' :: 'MonadReader' s m             => 'Getter' s a           -> m a
+-- 'view' :: ('MonadReader' s m, 'Monoid' a) => 'Control.Lens.Fold.Fold' s a             -> m a
+-- 'view' :: 'MonadReader' s m             => 'Control.Lens.Type.Simple' 'Control.Lens.Iso.Iso' s a       -> m a
+-- 'view' :: 'MonadReader' s m             => 'Control.Lens.Type.Simple' 'Control.Lens.Type.Lens' s a      -> m a
+-- 'view' :: ('MonadReader' s m, 'Monoid' a) => 'Control.Lens.Type.Simple' 'Control.Lens.Traversal.Traversal' s a -> m a
+-- @
+view :: MonadReader s m => Getting a s t a b -> m a
+view l = Reader.asks (runAccessor# (l Accessor))
 {-# INLINE view #-}
 
 -- | View the value of a 'Getter', 'Control.Lens.Iso.Iso',
@@ -202,6 +233,9 @@ view l = runAccessor# (l Accessor)
 -- >>> views _2 length (1,"hello")
 -- 5
 --
+-- As @views@ is commonly used to access the target of a 'Getter' or obtain a monoidal summary of the targets of a 'Fold',
+-- It may be useful to think of it as having one of these more restrictive signatures:
+--
 -- @
 -- 'views' ::             'Getter' s a             -> (a -> r) -> s -> r
 -- 'views' :: 'Monoid' m => 'Control.Lens.Fold.Fold' s a               -> (a -> m) -> s -> m
@@ -209,8 +243,18 @@ view l = runAccessor# (l Accessor)
 -- 'views' ::             'Control.Lens.Type.Simple' 'Control.Lens.Type.Lens' s a        -> (a -> r) -> s -> r
 -- 'views' :: 'Monoid' m => 'Control.Lens.Type.Simple' 'Control.Lens.Traversal.Traversal' s a   -> (a -> m) -> s -> m
 -- @
-views :: Getting r s t a b -> (a -> r) -> s -> r
-views l f = runAccessor# (l (accessor# f))
+--
+-- In a more general setting, such as when working with a monad transformer stack you can use:
+--
+-- @
+-- 'view' :: 'MonadReader' s m             => 'Getter' s a           -> m a
+-- 'view' :: ('MonadReader' s m, 'Monoid' a) => 'Control.Lens.Fold.Fold' s a             -> m a
+-- 'view' :: 'MonadReader' s m             => 'Control.Lens.Type.Simple' 'Control.Lens.Iso.Iso' s a       -> m a
+-- 'view' :: 'MonadReader' s m             => 'Control.Lens.Type.Simple' 'Control.Lens.Type.Lens' s a      -> m a
+-- 'view' :: ('MonadReader' s m, 'Monoid' a) => 'Control.Lens.Type.Simple' 'Control.Lens.Traversal.Traversal' s a -> m a
+-- @
+views :: MonadReader s m => Getting r s t a b -> (a -> r) -> m r
+views l f = Reader.asks (runAccessor# (l (accessor# f)))
 {-# INLINE views #-}
 
 -- | View the value pointed to by a 'Getter', 'Control.Lens.Iso.Iso' or
@@ -273,6 +317,9 @@ s ^. l = runAccessor (l Accessor s)
 -- 'Control.Lens.Fold.Fold' or 'Control.Lens.Traversal.Traversal' that points
 -- to a monoidal value.
 --
+-- >>> evalState ("hello","world") (use _1)
+-- "hello"
+--
 -- @
 -- 'use' :: 'MonadState' s m             => 'Getter' s a             -> m a
 -- 'use' :: ('MonadState' s m, 'Monoid' r) => 'Control.Lens.Fold.Fold' s r               -> m r
@@ -289,6 +336,9 @@ use l = State.gets (view l)
 -- 'Getter' in the current state, or use a summary of a
 -- 'Control.Lens.Fold.Fold' or 'Control.Lens.Traversal.Traversal' that
 -- points to a monoidal value.
+--
+-- >>> evalState ("hello","world") (uses _1 length)
+-- 5
 --
 -- @
 -- 'uses' :: 'MonadState' s m             => 'Getter' s a           -> (a -> r) -> m r
@@ -321,6 +371,7 @@ uses l f = State.gets (views l f)
 peruse :: MonadReader s m => Getting a s t a b -> m a
 peruse l = Reader.asks (^.l)
 {-# INLINE peruse #-}
+{-# DEPRECATED peruse "Use view instead. This will be removed in 3.8" #-}
 
 -- |
 -- Use the target of a 'Control.Lens.Type.Lens', 'Control.Lens.Iso.Iso' or
@@ -338,6 +389,7 @@ peruse l = Reader.asks (^.l)
 peruses :: MonadReader s m => Getting r s t a b -> (a -> r) -> m r
 peruses l f = Reader.asks (views l f)
 {-# INLINE peruses #-}
+{-# DEPRECATED peruses "Use views instead. This will be removed in 3.8" #-}
 
 -- | Useful for storing getters in containers.
 newtype ReifiedGetter s a = ReifyGetter { reflectGetter :: Getter s a }